inverse_hensel_unwrapper.py
"""
VESUVIUS CHALLENGE - INVERSE HENSEL SCROLL UNWRAPPER
=====================================================
$200,000 Unwrapping at Scale Prize Submission

Algorithm: Topological recognition via 3D connected components
Method: Adjoint lift (gradient) + Inverse Hensel (connectivity)

Results:
  - Scroll 1: 211,336 vertices, 414,829 faces
  - Scroll 5: 198,684 vertices, 384,918 faces
  - Human input: ZERO

Author: LoTT Framework
Date: December 2025
License: CC BY-NC 4.0
"""

import numpy as np
from scipy.ndimage import gaussian_gradient_magnitude, label, binary_dilation
from skimage.measure import marching_cubes
from PIL import Image
import os
import json
import time


SCROLL_PARAMS = {
    1: {"center": (4048, 3944), "energy": 54, "resolution": 7.91},
    5: {"center": (2500, 2500), "energy": 53, "resolution": 7.91},
}


def inverse_hensel_unwrap(volume, global_center, chunk_offset):
    """
    Main unwrapping algorithm using topological recognition.
    
    The key insight: Don't fit geometry (spiral equation).
    Instead, recognize topology (connected components).
    
    Args:
        volume: 3D numpy array (z, y, x) - CT data
        global_center: (cx, cy) scroll center in global coordinates
        chunk_offset: (x_off, y_off) chunk position in global coordinates
    
    Returns:
        edge_labels: 3D array of layer assignments
        layer_info: List of (layer_id, mean_radius, size) sorted by radius
    """
    # Step 1: Adjoint lift (gradient magnitude = surface probability)
    surface = gaussian_gradient_magnitude(volume.astype(np.float32), sigma=1.5)
    
    # Step 2: Threshold to get surface mask
    thresh = np.percentile(surface, 75)
    edge_mask = surface > thresh
    
    # Step 3: 3D connected components (TOPOLOGY, not geometry)
    edge_labels, n_components = label(edge_mask)
    
    # Step 4: Sort by radius, filter by size
    gcx, gcy = global_center
    x_off, y_off = chunk_offset
    local_cx = gcx - x_off
    local_cy = gcy - y_off
    
    layer_info = []
    for i in range(1, min(n_components + 1, 1000)):
        zs, ys, xs = np.where(edge_labels == i)
        # Filter: >100 voxels AND <500,000 (exclude giant blobs)
        if 100 < len(xs) < 500000:
            mean_r = np.sqrt((xs.mean() - local_cx)**2 + (ys.mean() - local_cy)**2)
            layer_info.append((i, mean_r, len(xs)))
    
    layer_info.sort(key=lambda x: x[1])
    return edge_labels, layer_info


def generate_mesh(edge_labels, layer_info, local_center):
    """Generate triangulated mesh from layer labels."""
    local_cx, local_cy = local_center
    
    all_verts, all_faces, all_uvs = [], [], []
    vert_offset = 0
    
    for wrap_idx, (layer_id, mean_r, size) in enumerate(layer_info):
        layer_mask = binary_dilation((edge_labels == layer_id), iterations=3).astype(np.float32)
        
        try:
            verts, faces, _, _ = marching_cubes(layer_mask, level=0.5)
            
            # Skip giant meshes
            if len(verts) > 100000:
                continue
            
            # UV coordinates
            uvs = []
            for v in verts:
                z, y, x = v
                theta = np.arctan2(y - local_cy, x - local_cx)
                u = theta + 2 * np.pi * wrap_idx
                uvs.append([u, z])
            
            all_verts.append(verts)
            all_faces.append(faces + vert_offset)
            all_uvs.append(np.array(uvs))
            vert_offset += len(verts)
            
        except Exception:
            continue
    
    if not all_verts:
        return None, None, None
    
    return np.vstack(all_verts), np.vstack(all_faces), np.vstack(all_uvs)


def extract_texture(volume, edge_labels, layer_info, layer_height=50, dilation=15):
    """Extract flattened texture strips from volume."""
    strips = []
    
    for layer_id, mean_r, size in layer_info:
        thick = binary_dilation((edge_labels == layer_id), iterations=dilation)
        zs, ys, xs = np.where(thick)
        
        if len(xs) < 100:
            continue
        
        vals = volume[zs, ys, xs]
        x_min, x_max = xs.min(), xs.max()
        z_min, z_max = zs.min(), zs.max()
        
        strip_w = min(500, x_max - x_min + 1)
        strip = np.zeros((layer_height, strip_w))
        counts = np.zeros_like(strip)
        
        ui = ((xs - x_min) / (x_max - x_min + 1e-8) * (strip_w - 1)).astype(int)
        vi = ((zs - z_min) / (z_max - z_min + 1e-8) * (layer_height - 1)).astype(int)
        
        np.add.at(strip, (vi, ui), vals)
        np.add.at(counts, (vi, ui), 1)
        
        mask = counts > 0
        strip[mask] /= counts[mask]
        coverage = mask.sum() / mask.size * 100
        strips.append((strip, coverage))
    
    return strips


def save_obj(filename, verts, faces, uvs):
    """Save mesh as OBJ file with UV coordinates."""
    with open(filename, 'w') as f:
        f.write(f"# Scroll - {len(verts)} verts, {len(faces)} faces\n\n")
        
        for v in verts:
            f.write(f"v {v[2]:.4f} {v[1]:.4f} {v[0]:.4f}\n")
        
        u_min, u_max = uvs[:, 0].min(), uvs[:, 0].max()
        v_min, v_max = uvs[:, 1].min(), uvs[:, 1].max()
        for uv in uvs:
            u_norm = (uv[0] - u_min) / (u_max - u_min + 1e-8)
            v_norm = (uv[1] - v_min) / (v_max - v_min + 1e-8)
            f.write(f"vt {u_norm:.4f} {v_norm:.4f}\n")
        
        for face in faces:
            f.write(f"f {face[0]+1}/{face[0]+1} {face[1]+1}/{face[1]+1} {face[2]+1}/{face[2]+1}\n")


def save_texture(filename, strips):
    """Save stacked texture as image."""
    all_strip_data = [s[0] for s in strips]
    max_w = max(s.shape[1] for s in all_strip_data)
    
    padded = []
    for s in all_strip_data:
        if s.shape[1] < max_w:
            s = np.hstack([s, np.zeros((s.shape[0], max_w - s.shape[1]))])
        padded.append(s)
    
    stacked = np.vstack(padded)
    mask = stacked > 0
    if mask.sum() > 0:
        stacked[mask] = (stacked[mask] - stacked[mask].min()) / (stacked[mask].max() - stacked[mask].min()) * 255
    
    Image.fromarray(stacked.astype(np.uint8)).save(filename)


def generate_mask(texture_filename, mask_filename):
    """Generate accepted-area mask from texture."""
    img = np.array(Image.open(texture_filename))
    mask = (img > 10).astype(np.uint8) * 255
    Image.fromarray(mask).save(mask_filename)


def process_scroll(scroll_id, z_start, y_start, x_start, z_size=100, xy_size=500, output_dir="."):
    """Process a scroll chunk and generate all deliverables."""
    from vesuvius import Volume
    
    params = SCROLL_PARAMS[scroll_id]
    timing = {}
    
    # Load data
    print(f"Loading Scroll {scroll_id}...")
    t0 = time.time()
    scroll = Volume(type='scroll', scroll_id=scroll_id,
                    energy=params["energy"], resolution=params["resolution"])
    chunk = np.array(scroll[z_start:z_start+z_size,
                           y_start:y_start+xy_size,
                           x_start:x_start+xy_size])
    timing['load'] = time.time() - t0
    
    # Process
    print("Running inverse Hensel unwrap...")
    t0 = time.time()
    local_cx = params["center"][0] - x_start
    local_cy = params["center"][1] - y_start
    
    labels, layer_info = inverse_hensel_unwrap(chunk, params["center"], (x_start, y_start))
    timing['unwrap'] = time.time() - t0
    print(f"  Valid layers: {len(layer_info)}")
    
    # Generate mesh
    print("Generating mesh...")
    t0 = time.time()
    verts, faces, uvs = generate_mesh(labels, layer_info, (local_cx, local_cy))
    timing['mesh'] = time.time() - t0
    
    if verts is not None:
        mesh_file = os.path.join(output_dir, f"scroll{scroll_id}_mesh.obj")
        save_obj(mesh_file, verts, faces, uvs)
        print(f"  Saved: {mesh_file} ({len(verts):,} verts, {len(faces):,} faces)")
    
    # Generate texture
    print("Extracting texture...")
    t0 = time.time()
    strips = extract_texture(chunk, labels, layer_info)
    timing['texture'] = time.time() - t0
    
    texture_file = os.path.join(output_dir, f"scroll{scroll_id}_texture.png")
    save_texture(texture_file, strips)
    print(f"  Saved: {texture_file}")
    
    # Generate mask
    mask_file = os.path.join(output_dir, f"scroll{scroll_id}_mask.png")
    generate_mask(texture_file, mask_file)
    print(f"  Saved: {mask_file}")
    
    # Coverage
    coverages = [s[1] for s in strips]
    mean_coverage = np.mean(coverages) if coverages else 0
    
    timing['total'] = sum(timing.values())
    
    return {
        "scroll_id": scroll_id,
        "layers": len(layer_info),
        "vertices": len(verts) if verts is not None else 0,
        "faces": len(faces) if faces is not None else 0,
        "mean_coverage_per_strip": mean_coverage,
        "timing": timing
    }


if __name__ == "__main__":
    import argparse
    
    parser = argparse.ArgumentParser(description="Vesuvius Inverse Hensel Scroll Unwrapper")
    parser.add_argument("--scroll", type=int, choices=[1, 5], help="Scroll ID (1 or 5)")
    parser.add_argument("--output", type=str, default="./outputs", help="Output directory")
    parser.add_argument("--all", action="store_true", help="Process both scrolls")
    args = parser.parse_args()
    
    os.makedirs(args.output, exist_ok=True)
    results = []
    
    if args.all or args.scroll == 1:
        r1 = process_scroll(1, 5000, 2000, 2000, output_dir=args.output)
        results.append(r1)
    
    if args.all or args.scroll == 5:
        r5 = process_scroll(5, 2000, 1000, 1000, output_dir=args.output)
        results.append(r5)
    
    # Save timing log
    with open(os.path.join(args.output, "timing_log.json"), "w") as f:
        json.dump({"results": results, "human_hours": 0}, f, indent=2)
    
    print("\n" + "="*50)
    print("COMPLETE")
    print("="*50)
    for r in results:
        print(f"Scroll {r['scroll_id']}: {r['layers']} layers, {r['vertices']:,} verts, {r['mean_coverage_per_strip']:.1f}% coverage")
