requirements.txt
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.21.0
pillow>=9.0.0
vesuvius>=0.1.0
